% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_nonwear.R
\name{add_nonwear}
\alias{add_nonwear}
\title{Convenience wrapper for \code{categorize_nonwear_troiano}}
\usage{
add_nonwear(
  acc_data,
  counts,
  algorithm = nonwear_troiano,
  ...,
  nonwear = "nonwear"
)
}
\description{
Convenience wrapper for \code{categorize_nonwear_troiano}
}
