% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_pa.R
\name{add_pa}
\alias{add_pa}
\title{Add PA categories to accelerometer data}
\usage{
add_pa(acc_data, counts, age, cut_params = pa_params_troiano, pa = "pa")
}
\description{
Convenience wrapper for \code{categorize_pa}
}
