% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pa.R
\name{add_pa_category}
\alias{add_pa_category}
\title{Add PA category to accelerometer data}
\usage{
add_pa_category(acc_data, age = NA, cut_params = pacuts_troiano,
  use_magnitude = FALSE, add_age = FALSE, pa = "pa")
}
\arguments{
\item{acc_data}{Accelerometer tsibble object.}

\item{age}{Age in years as a numeric value.}

\item{cut_params}{Cut parameters, default is Troiano cut parameters.}

\item{use_magnitude}{Should magnitude be used instead of axis1 (the default)?}

\item{add_age}{Should the age value be appended as a separate column to the
output? Can accept a string value, which will be used as the new age column
name.}

\item{pa}{The name for the new column containing PA category values.}
}
\value{
The original accelerometer data with an appended column for physical
activity category and optionally another appended column for age.
ActiGraph settings are stored as a tibble in the settings attribute.
}
\description{
Adds PA category as a new column in accelerometer data. Returns original
data with appended column(s).
}
