% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pa.R
\name{calculate_pa_category}
\alias{calculate_pa_category}
\title{Calculate PA category}
\usage{
calculate_pa_category(x, epoch_len, age, cut_params = pacuts_troiano)
}
\arguments{
\item{x}{Numeric vector of counts.}

\item{epoch_len}{Epoch length in seconds.}

\item{age}{Age in years as a numeric value.}

\item{cut_params}{Cut parameters, default is Troiano cut parameters.}
}
\value{
An ordinal vector of the same length as \code{x}.
}
\description{
Calculates PA category given count data and other paramters. Returns a
vector of values.
}
