% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_nonwear.R
\name{categorize_nonwear}
\alias{categorize_nonwear}
\title{Categorize nonwear}
\usage{
categorize_nonwear(counts, epoch_len, method = nonwear_troiano, ...)
}
\arguments{
\item{counts}{Numeric vector of accelerometer values.}

\item{epoch_len}{Epoch length in seconds.}
}
\description{
Categorize nonwear
}
