% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_timezone}
\alias{change_timezone}
\title{Get timezone of accelerometer data}
\usage{
change_timezone(acc_data, tz = Sys.timezone())
}
\arguments{
\item{acc_data}{Accelerometer tsibble object.}

\item{tz}{Timezone to change to.}
}
\value{
Accelerometer data with updated timezone.
}
\description{
Get timezone of accelerometer data
}
