% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_epochlength}
\alias{get_epochlength}
\title{Get epoch length of accelerometer data}
\usage{
get_epochlength(acc_data)
}
\arguments{
\item{acc_data}{Accelerometer tsibble object.}
}
\value{
The epoch length in seconds as an integer value.
}
\description{
Get epoch length of accelerometer data
}
