% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_acc.R
\name{get_setting}
\alias{get_setting}
\title{Get settings from accelerometer data}
\usage{
get_setting(acc_data, setting)
}
\arguments{
\item{acc_data}{Accelerometer tsibble object.}

\item{setting}{Setting to extract as a string.}
}
\value{
The value of the stored setting. Most of the time an atomic vector of length 1.
}
\description{
Get settings from accelerometer data
}
