% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_acc.R
\name{get_timezone}
\alias{get_timezone}
\title{Get timezone of accelerometer data}
\usage{
get_timezone(acc_data)
}
\arguments{
\item{acc_data}{Accelerometer tsibble object.}
}
\value{
The timezone as a string value.
}
\description{
Get timezone of accelerometer data
}
