% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_acc.R
\name{list_settings}
\alias{list_settings}
\title{List settings from accelerometer data}
\usage{
list_settings(acc_data)
}
\arguments{
\item{acc_data}{Accelerometer tsibble object.}
}
\value{
A list of stored setting names as a string vector.
}
\description{
List settings from accelerometer data
}
