% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_pa.R
\name{pa_params_matthews}
\alias{pa_params_matthews}
\title{Cut parameters for Matthews}
\usage{
pa_params_matthews(age = 18)
}
\arguments{
\item{age}{Age in years as a numeric value.}
}
\value{
A list of parameters for the cut function.
}
\description{
Generates cut parameters based on Matthews values for adults.
}
