% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pa.R
\name{pacuts_troiano}
\alias{pacuts_troiano}
\title{Cut parameters for Troiano}
\usage{
pacuts_troiano(age)
}
\arguments{
\item{age}{Age in years as a numeric value.}
}
\value{
A list of parameters for the cut function.
}
\description{
Generates cut parameters based on Troiano values given an age.
}
