% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_actigraph.R
\name{read_csv_actigraph}
\alias{read_csv_actigraph}
\title{Read accelerometer data from an *.csv file}
\usage{
read_csv_actigraph(file, tz = "UTC", preamble = FALSE)
}
\arguments{
\item{file}{Full path to a csv file to read.}

\item{tz}{Time zone for the output data, converted from UTC (the default).}

\item{preamble}{Keep csv file preamble as an attribute?}
}
\value{
A \code{tsibble} (\code{tbl_ts}) of accelerometer data with at least two columns: timestamp and axis1.
}
\description{
Read ActiGraph accelerometer data from a database stored in an CSV file. Return a tsibble.
}
