% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_pa.R
\name{categorize_pa}
\alias{categorize_pa}
\title{Categorize accelerometer epochs into PA categories}
\usage{
categorize_pa(counts, epoch_len, age, cut_params = pacuts_troiano)
}
\arguments{
\item{counts}{Numeric vector of accelerometer values.}

\item{epoch_len}{Epoch length in seconds.}

\item{age}{Age in years as a numeric value.}

\item{cut_params}{Cut parameters, default is Troiano cut parameters.}
}
\value{
An ordinal vector of the same length as \code{x}.
}
\description{
Calculates PA category given count data and other paramters. Returns a
vector of values.
}
