% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{read_agd}
\alias{read_agd}
\title{Read accelerometer data from an *.agd file}
\usage{
read_agd(file, tz = "UTC")
}
\arguments{
\item{file}{Full path to an agd file to read.}

\item{tz}{Time zone for the output data, converted from UTC (the default).}
}
\value{
A \code{tsibble} (\code{tbl_ts}) of accelerometer data with at least two columns: timestamp and axis1.
ActiGraph settings are stored as a tibble in the settings attribute.
}
\description{
Read ActiGraph accelerometer data from a database stored in an AGD file. Return a tsibble.
}
